+++ src/providers/openai/chat.ts
